refactored code and tests to use a Weather class
