+++ spec/airport_spec.rb
@@ -5,7 +5,7 @@ describe Airport do
   describe '#land' do
     it { is_expected.to respond_to(:land).with(1).argument }
 
-    it 'returns a success message when we land a plane' do
+    it 'should return a success message when we land a plane' do
       allow(subject).to receive(:current_weather) { "Sunny" }
       expect(subject.land(Plane.new)).to eq "Plane landed"
     end
@@ -27,36 +27,34 @@ describe Airport do
       subject.land(my_plane)
       expect { subject.land(my_plane) }.to raise_error "That plane is already in the airport"
     end
-
   end
 
   describe '#plane_list' do
     it { is_expected.to respond_to :plane_list }
 
-    it 'returns no planes when airport is empty' do
+    it 'should return no planes when airport is empty' do
       expect(subject.plane_list).to eq []
     end
 
-    it 'returns a plane when I check what has landed' do
+    it 'should show that my landed plane is in the airport' do
       allow(subject).to receive(:current_weather) { "Sunny" }
       my_plane = Plane.new
       subject.land(my_plane)
       expect(subject.plane_list).to include my_plane
     end
-
   end
 
   describe '#take_off' do
     it { is_expected.to respond_to(:take_off).with(1).argument }
 
-    it 'returns a confirmation that the plane has taken off' do
+    it 'should return a confirmation that the plane has taken off' do
       allow(subject).to receive(:current_weather) { "Sunny" }
       my_plane = Plane.new
       subject.land(my_plane)
       expect(subject.take_off(my_plane)).to eq "Successful take off"
     end
 
-    it 'raises an error if I try to make a non-existent plane take off' do
+    it 'should raise an error if I try to make a non-existent plane take off' do
       allow(subject).to receive(:current_weather) { "Sunny" }
       my_plane = Plane.new
       subject.land(my_plane)
@@ -78,16 +76,15 @@ describe Airport do
       allow(subject).to receive(:current_weather) { "Stormy" }
       expect { subject.take_off(my_plane) }.to raise_error "No take-off allowed due to the weather"
     end
-
   end
 
-  describe '#current_weather' do
-    it { is_expected.to respond_to(:current_weather) }
-
-    it 'should return either Sunny or Stormy' do
-      expect(subject.current_weather).to eq("Sunny").or eq("Stormy")
-    end
-  end
+  # describe '#current_weather' do
+  #   it { is_expected.to respond_to(:current_weather) }
+  #
+  #   it 'should return either Sunny or Stormy' do
+  #     expect(subject.current_weather).to eq("Sunny").or eq("Stormy")
+  #   end
+  # end
 
   describe '#capacity' do
     it { is_expected.to respond_to(:capacity) }
